+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Our versioning strategy is as follows:
 ### 🎉 New Features & Improvements
 
 * `[core]` [DesignLibrary] Call partial layout rendering endpoint via Envoy and ContextID ([#111](https://github.com/Sitecore/content-sdk/pull/111))
+* `[core]` `[nextjs]` [DesignLibrary] Include metadata in the Design Library rendering mechanism ([#118](https://github.com/Sitecore/content-sdk/pull/118))
 
 ### 🛠 Breaking Changes